Extract rewriting and CB.
